@@ -162,8 +164,10 @@ IF (MSVC)
     SET(LIQUID_DLL "${PROJECT_SOURCE_DIR}/external/liquid-dsp/msvc/${EX_PLATFORM}/libliquid.dll" CACHE STRING "Liquid-DSP DLL")
     SET(HAMLIB_DLLS "${PROJECT_SOURCE_DIR}/external/hamlib/${EX_PLATFORM}/libhamlib-4.dll" CACHE STRING "HAMLIB DLLS")
 ELSE (MSVC)
+    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++20")
+
     ADD_DEFINITIONS(
-        -std=c++0x 
+        # -std=c++0x 
         -pthread
     )
 ENDIF(MSVC)
@@ -742,9 +774,11 @@ IF (APPLE AND BUNDLE_APP)
         )
         ENDIF()
     ENDIF()
-    
+
+    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++20")
+
     ADD_DEFINITIONS(
-        -std=c++0x 
+        # -std=c++0x 
         -pthread
         -D_OSX_APP_
     )   
